universal_skill_auto/requirements.txt
requests
beautifulsoup4
newspaper3k
pyalex